Extract colors and fonts to css vars
