+++ src/App.css
@@ -1,5 +1,5 @@
 html {
-    background-color: #0b1d26;
+    background-color: var(--color-bg);
     min-height: 2000px;
     font-size: 100%;
 }
