+++ src/components/navbar/NavBarEntry.css
@@ -1,6 +1,6 @@
 .navbar__entry {
     flex-grow: 0;
-    font-family: Amiko;
+    font-family: var(--font-primary);
     font-size: 16px;
     font-weight: bold;
     color: #fff;
