+++ src/components/story/StoryEntry.css
@@ -25,7 +25,7 @@
     height: 240px;
     flex-grow: 0;
     opacity: 0.1;
-    font-family: Amiko;
+    font-family: var(--font-primary);
     font-size: 240px;
     font-weight: bold;
     font-stretch: normal;
@@ -58,7 +58,7 @@
 .subtitle__text {
     height: 24px;
     flex-grow: 1;
-    font-family: Amiko;
+    font-family: var(--font-primary);
     font-size: 18px;
     font-weight: bold;
     font-stretch: normal;
@@ -73,7 +73,7 @@
     height: 76px;
     align-self: stretch;
     flex-grow: 0;
-    font-family: AbhayaLibre;
+    font-family: var(--font-secondary);
     font-size: 64px;
     font-weight: normal;
     font-stretch: normal;
@@ -88,7 +88,7 @@
     height: 160px;
     align-self: stretch;
     flex-grow: 0;
-    font-family: Amiko;
+    font-family: var(--font-primary);
     font-size: 18px;
     font-weight: bold;
     font-stretch: normal;
@@ -104,7 +104,7 @@
     height: 24px;
     flex-grow: 0;
     margin: 0 4px 0 0;
-    font-family: Amiko;
+    font-family: var(--font-primary);
     font-size: 18px;
     font-weight: bold;
     font-stretch: normal;
