+++ src/components/story/StoryHeader.css
@@ -1,4 +1,5 @@
 .story-header {
+    z-index: -2;
     position: fixed;
     width: 60%;
     max-width: 1000px;
@@ -17,7 +18,7 @@
     height: 208px;
     align-self: stretch;
     flex-grow: 0;
-    font-family: Abhaya Libre;
+    font-family: var(--font-secondary);
     font-size: 88px;
     font-weight: normal;
     font-stretch: normal;
@@ -51,7 +52,7 @@
     text-transform: uppercase;
     height: 24px;
     flex-grow: 1;
-    font-family: Amiko;
+    font-family: var(--font-primary);
     font-size: 18px;
     font-weight: bold;
     font-stretch: normal;
