+++ src/components/story/StoryLine.css
@@ -4,6 +4,5 @@
     justify-content: center;
     align-items: stretch;
     gap: 200px;
-    margin: 440px 229px 469px;
-    padding: 64px 0;
+    padding: 1200px 0;
 }
