+++ src/containers/story/Story.css
@@ -1,5 +1,5 @@
 .story__bg-clouds {
-    z-index: -1;
+    z-index: -2;
     position: absolute;
     width: 100vw;
     height: 1100px;
@@ -7,7 +7,7 @@
 }
 
 .story__mg-mountains {
-    z-index: -1;
+    z-index: -2;
     position: absolute;
     width: 100vw;
     top: 53%;
@@ -16,7 +16,7 @@
 }
 
 .story__fg-hill {
-    z-index: 1;
+    z-index: -1;
     position: absolute;
     width: 100vw;
     top: 48%;
@@ -25,7 +25,7 @@
 }
 
 .story__gradient-top {
-    z-index: -1;
+    z-index: -2;
     position: absolute;
     width: 100vw;
     height: 600px;
