+++ src/index.css
@@ -11,3 +11,9 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+:root {
+  --color-bg: #0b1d26;
+  --font-primary: Amiko;
+  --font-secondary: Abhaya Libre;
+}
